✅ Update: Migrate to Integer IDs & improve data model 📄🚀

+++ src/main/java/com/ince01/todo_server/application/port/in/TodoService.java
@@ -5,13 +5,13 @@ import com.ince01.todo_server.domain.model.Todo;
 import java.util.List;
 
 public interface TodoService {
-    Todo createTodo(Long userId, String title, String description);
+    Todo createTodo(Integer userId, String title, String description);
 
-    Todo getTodoById(Long userId, Long todoId);
+    Todo getTodoById(Integer userId, Integer todoId);
 
-    List<Todo> getAllTodos(Long userId);
+    List<Todo> getAllTodos(Integer userId);
 
-    Todo updateTodo(Long userId, Long todoId, String title, String description, boolean completed);
+    Todo updateTodo(Integer userId, Integer todoId, String title, String description, boolean completed);
 
-    void deleteTodo(Long userId, Long todoId);
+    void deleteTodo(Integer userId, Integer todoId);
 }
