+++ src/main/java/com/ince01/todo_server/domain/exception/TodoNotFoundException.java
@@ -1,7 +1,7 @@
 package com.ince01.todo_server.domain.exception;
 
 public class TodoNotFoundException extends RuntimeException {
-    public TodoNotFoundException(Long id) {
+    public TodoNotFoundException(Integer id) {
         super("Todo with ID " + id + " not found");
     }
 }
